# docs/swagger.yaml
basePath: /
definitions:
  handlers.AuthResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expires_in:
        example: 3600
        type: integer
      message:
        example: Login successful
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/handlers.UserResponse'
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        example: Invalid input
        type: string
    type: object
  handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
      two_factor_token:
        type: string
    required:
    - email
    - password
    type: object
...
  handlers.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  handlers.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  handlers.RequestPasswordResetRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  handlers.ResendVerificationRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  handlers.SuccessResponse:
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
...
  handlers.TwoFactorSetupResponse:
    properties:
      backup_codes:
        example:
        - "123456"
        - "789012"
        - "345678"
        items:
          type: string
        type: array
      qr_code_url:
        example: data:image/png;base64,iVBOR...
        type: string
      secret_key:
        example: JBSWY3DPEHPK3PXP
        type: string
    type: object
...
  handlers.UserResponse:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      email_verified:
        example: true
        type: boolean
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      last_login_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      name:
        example: John Doe
        type: string
      two_factor_enabled:
        example: false
        type: boolean
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
  handlers.VerifyEmailRequest:
    properties:
      token:
        type: string
    required:
    - token
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive task management API with real-time collaboration features
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Real Task API
  version: "1.0"
paths:
...
  /api/auth/2fa/setup:
    post:
      consumes:
      - application/json
      description: Initialize two-factor authentication setup and get QR code
      produces:
      - application/json
      responses:
        "200":
          description: 2FA setup initiated successfully
          schema:
            $ref: '#/definitions/handlers.TwoFactorSetupResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Setup two-factor authentication
      tags:
      - Two-Factor Authentication
...
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: Generate new access token using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: New access token generated
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Refresh access token
      tags:
      - Authentication
  /api/auth/request-password-reset:
    post:
      consumes:
      - application/json
      description: Send password reset link to user's email address
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RequestPasswordResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset email sent
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Invalid email or user not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Request password reset
      tags:
      - Authentication
  /api/auth/resend-verification:
    post:
      consumes:
      - application/json
      description: Resend email verification link to user's email address
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ResendVerificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Verification email sent
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Invalid email or user not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Resend verification email
      tags:
      - Authentication
  /api/auth/verify-email:
    get:
      consumes:
      - application/json
      description: Verify user email address using verification token
      parameters:
      - description: Verification token (GET method)
        in: query
        name: token
        type: string
      - description: Verification token (POST method)
        in: body
        name: request
        schema:
          $ref: '#/definitions/handlers.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Email verified successfully
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Invalid or expired token
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Verify user email
      tags:
      - Authentication
    post:
      consumes:
      - application/json
      description: Verify user email address using verification token
      parameters:
      - description: Verification token (GET method)
        in: query
        name: token
        type: string
      - description: Verification token (POST method)
        in: body
        name: request
        schema:
          $ref: '#/definitions/handlers.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Email verified successfully
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Invalid or expired token
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Verify user email
      tags:
      - Authentication
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password, supports 2FA
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/handlers.AuthResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Authentication failed
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "423":
          description: Account locked
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: User login
      tags:
      - Authentication
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user account with email verification
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Register a new user
      tags:
      - Authentication
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"